pyproject.toml
[project]
name = "hr-data-cleansing-demo-app"
version = "1.0.0"
description = "Production-grade HR analytics and data cleaning system."
authors = [
    {name = "Akira Koshimizu", email = "882akira@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "streamlit>=1.24.0",
    "plotly>=5.15.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.0",
    "anthropic>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "flake8>=6.0.0",
    "black>=23.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
